added basic powerup functionality and pickup capability

+++ Assets/Scripts/PlayerControl.cs
@@ -6,8 +6,8 @@ public class PlayerControl : MonoBehaviour {
 
 	static public PlayerControl S;
 
-	public int speed = 10;
-	public int jumpPower = 10;
+	public float speed = 10;
+	public float jumpPower = 10;
 
 	private Rigidbody rb;
